requirements_test.txt
--find-links http://pip.lpds.sztaki.hu/packages
--no-index

argparse==1.2.1
cov-core==1.15.0
coverage==3.7.1
dateutils==0.6.6
pika==0.9.14
python-dateutil==2.2
pytz==2014.9
PyYAML==3.11
six==1.8.0
wsgiref==0.1.2
nose==1.3.4
nose-cov==1.6
requests==2.4.3
-e .
